+++ CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Next Release
 
+## 0.0.25
+
+- [router] Fix bug in subgraph loop.
+
+## 0.0.24
+
+- [sdk] Use local clock for all time calcs.
+
 ## 0.0.22
 
 - [sdk] Add back `connectMessaging` with proper functionality.